added variable size functionality to benchmark.sh

+++ benchmark.sh
@@ -1,20 +1,27 @@
 #!/bin/bash
 
-export PATH=.:$PATH
+TESTSIZE="$1"
+if [ -z "$TESTSIZE" ]; then
+	TESTSIZE="1M"
+fi
 
-HASHES="md5 sha1 sha224 sha256 sha384 sha512"
+if [ -n "$(which pv)" ]; then
+	export PATH=.:$PATH
 
-TEMPFILE="sparse-$RANDOM.file"
-dd if=/dev/zero of=$TEMPFILE bs=1k count=0 seek=1M 2> /dev/null
+	HASHES="md5 sha1 sha224 sha256 sha384 sha512"
 
-for HASH in $(echo $HASHES | tr ' ' '\n'); do
-	printf "\n$HASH\n"
-	for POSTFIX in $(echo "sum pass" | tr ' ' '\n'); do
-		pv -cN "$HASH$POSTFIX" < $TEMPFILE | $HASH$POSTFIX &> /dev/null
+	TEMPFILE="sparse-$RANDOM.file"
+	dd if=/dev/zero of=$TEMPFILE bs=1 count=0 seek=$TESTSIZE 2> /dev/null
+
+	for HASH in $(echo $HASHES | tr ' ' '\n'); do
+		printf "\n$HASH\n"
+		for POSTFIX in $(echo "sum pass" | tr ' ' '\n'); do
+			pv -cN "$HASH$POSTFIX" < $TEMPFILE | $HASH$POSTFIX &> /dev/null
+		done
+		for POSTFIX in $(echo "pass sum" | tr ' ' '\n'); do
+			pv -cN "$HASH$POSTFIX" < $TEMPFILE | $HASH$POSTFIX &> /dev/null
+		done
 	done
-	for POSTFIX in $(echo "pass sum" | tr ' ' '\n'); do
-		pv -cN "$HASH$POSTFIX" < $TEMPFILE | $HASH$POSTFIX &> /dev/null
-	done
-done
 
-rm -f $TEMPFILE
+	rm -f $TEMPFILE
+fi
